Rename to NoUnneededImportAliasFixer, supports comments, minor fixes after review
